requirements.txt
torch >= 1.3.0
torchvision >= 0.6.0

pillow~=8.0.1
numpy~=1.19.2
einops~=0.3.0
setuptools~=50.3.1
apex~=0.1